// sourcecode/BTrak.Packages/snova-app-builder-creation-components/app-builder-module-components/app-builder-components/src/lib/builder/models/custom-query-headers.model.ts
export class CustomQueryHeadersModel {
    field: string;
    title: string;
    titleField: string;
    hidden: any;
    filter: string;
    maxLength: number;
    orderIndex: number;
    width: any;
    isNullable: boolean;
    includeInFilters: boolean;
    isAvailableForFiltering: boolean;
    subQuery?: string;
    subQueryType?: string;
    subQueryTypeId?: string;
    columnFormatTypeId?:string;
    columnFormatTypes:any;
    formattype:any;
    columnAltName:string;
    columnFormatType:string;
    columnFontColor: string;
    rowBackgroundColor: string;
    headerBackgroundColor: string;
    headerFontColor: string;
    columnFontFamily: string;
    columnBackgroundColor: string;
    isCurrency:string;
}